requirements.txt
Django
djangorestframework
drf_util